+++ src/main/java/org/tallerjava/moduloMediosDePago/dominio/Tarifa.java
@@ -1,20 +1,20 @@
 package org.tallerjava.moduloMediosDePago.dominio;
 
 public class Tarifa {
-    private int monto;
+    private double monto;
 
-    public Tarifa(int monto) {
+    public Tarifa(double monto) {
         this.monto = monto;
     }
 
     public Tarifa() {
     }
 
-    public int getMonto() {
+    public double getMonto() {
         return monto;
     }
 
-    public void setMonto(int monto) {
+    public void setMonto(double monto) {
         this.monto = monto;
     }
 }
