+++ src/main/java/org/tallerjava/moduloMediosDePago/dominio/TarifaComun.java
@@ -1,7 +1,7 @@
 package org.tallerjava.moduloMediosDePago.dominio;
 
 public class TarifaComun extends Tarifa {
-    public TarifaComun(int monto) {
+    public TarifaComun(double monto) {
         super(monto);
     }
